% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EWR_functions.R
\name{bioProp_EWR_COVfactor_adjustment}
\alias{bioProp_EWR_COVfactor_adjustment}
\title{Biological proportion varaince computation}
\usage{
bioProp_EWR_COVfactor_adjustment(
  X,
  Y,
  Z,
  X_padding = 1,
  PC_control_num = 10,
  bootstrap_num = 20,
  bootstrap_samples_flg = F,
  number_XZ_factor_model = 1,
  covariate_Z_factor = 1
)
}
\arguments{
\item{X}{a vector which denote the levels of the target protein.}

\item{Y}{a vector (same length of X) which denote the levels of the helper protein, should be significantly correlated with X.}

\item{Z}{a matrix (each row correspond to one element in X and Y) of auxiliary proteins that tags biological variance in X. We recommend at least 100 columns.}

\item{X_padding}{default 1: increasing this number will putting more weights on target protein using a padding procedure. Usually not required as factor model is already controlling latent factors in Z.}

\item{PC_control_num}{default 10: the number of PCs in Z that would be regressed out from Z.}

\item{bootstrap_num}{Default 20. Number of Bootstrapping or Jackknifing blocks used to compute the confidence interval of the estimate.}

\item{bootstrap_samples_flg}{Default F: when TRUE, bootstrapping across individuals; recommend FALSE where sampling is done across the Z columns, this is more calibrated as EWR is estimated across columns of Zs.}

\item{number_XZ_factor_model}{Default 1: the choice of latent factors of X that are independent of Y.}

\item{covariate_Z_factor}{Default 1: The choice of optimal latent factors of Z that are independent of X and Y; use one factor to control the latent structure in Z.}
}
\value{
a list:
Bioprop_xi1_xi2: estimate of biological proportion (i.e. 1-noise) variance in protein level X.
sampling_se: standard error of the estimate Bioprop_xi1_xi2; estimated through sampling procedure.
sampling_mean: mean across the sample estimate (could be used as a bias detection tool from Jackknifing).
PC_variance: the proportion of X variance tagged by PCs of Z, which are regressed out from Z and then added back to the final estimate.
PxPy
PxPy_se: standard error of PxPy estimate.
beta_EWR: Slope of covariance y2z on x2z.
beta_EWR_se: standard error of beta_EWR estimate.
approx_Bioprop_xi1_xi2: term A in the paper.
ratio_2: term R2 in the paper.
ratio_1: term R2=1 in the paper.
sigma_X: residual variance in X that is not captured by the latent factors.
sigma_Y: residual variance in Y that is not captured by the latent factors.
itr_quasi_newton: the number of iteration between the (i) quasi-newton MLE and (ii) rotation of factor loadings. See the paper supplementary for details.
}
\description{
Biological proportion varaince computation
}
\examples{
library(ExpressionWideRegression)
full_run_results_big_wrapper <- EWR(X_EXAMPLE,Z_EXAMPLE)

}
