% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EWR_functions.R
\name{EWR}
\alias{EWR}
\title{Estimate non-noise proportion of variance in the protein measurement, using a set of auxiliary proteins.}
\usage{
EWR(
  target_protein,
  auxiliary_proteins,
  use_as_helper = 10,
  Factor_Model_selection = T,
  XZ_factor = 1,
  Z_factor = 0
)
}
\arguments{
\item{target_protein:}{a vector which denote the levels of the target protein.}

\item{auxiliary_proteins:}{matrix (each row correspond to one element in target proteins) of auxiliary proteins that tags biological variance in target proteins. We recommend at least 100 columns.}

\item{use_as_helper:}{default 10: we select top #use_as_helper strongly associated columns in auxiliary_proteins as helper proteins for the estimation. Each estimate should be estimate the same value.}

\item{Factor_Model_selection:}{default TRUE: true means performing a model selection to choose the model structure. If FALSE, default or user specified values will be used.}

\item{XZ_factor:}{Default 1: the choice of latent factors of X that are independent of Y.}

\item{Z_factor:}{Default 0: the choice of latent factors of Z that are independent of X and Y.}
}
\value{
list:
Bioprop_xi1_xi2: estimate of biological proportion (i.e. 1-noise) variance in protein level X.
sampling_se: standard error of the estimate Bioprop_xi1_xi2; estimated through sampling procedure.
sampling_mean: mean across the sample estimate (could be used as a bias detection tool from Jackknifing).
PC_variance: the proportion of X variance tagged by PCs of Z, which are regressed out from Z and then added back to the final estimate.
PxPy
PxPy_se: standard error of PxPy estimate.
beta_EWR: Slope of covariance y2z on x2z.
beta_EWR_se: standard error of beta_EWR estimate.
approx_Bioprop_xi1_xi2: term A in the paper.
ratio_2: term R2 in the paper.
ratio_1: term R2=1 in the paper.
sigma_X: residual variance in X that is not captured by the latent factors.
sigma_Y: residual variance in Y that is not captured by the latent factors.
itr_quasi_newton: the number of iteration between the (i) quasi-newton MLE and (ii) rotation of factor loadings. See the paper supplementary for details.
Y_idx: the column index of auxiliary_proteins that are chosen as the helper protein.
XZ_factor: the choice of latent factors of X that are independent of Y.
Z_factor: the choice of latent factors of Z that are independent of X and Y.
}
\description{
Estimate non-noise proportion of variance in the protein measurement, using a set of auxiliary proteins.
}
\examples{
full_run_results_big_wrapper <- EWR(X_EXAMPLE,Z_EXAMPLE, use_as_helper = 2)
}
