% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EXAMPLE_COV}
\alias{EXAMPLE_COV}
\title{Example of covariance of data matrix -- the factor model is based on the covariance matrix.}
\format{
\subsection{\code{EXAMPLE_COV}}{

A matrix with 200 rows and 200 columns.
}
}
\usage{
EXAMPLE_COV
}
\description{
The data is generated by sampling procedure for showing the step-to-step application of the function.
}
\keyword{datasets}
