% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionWideRegression-package.R
\docType{package}
\name{ExpressionWideRegression-package}
\alias{ExpressionWideRegression}
\alias{ExpressionWideRegression-package}
\title{ExpressionWideRegression: Estimate biologically meaningful variance and measurement noise of protein level.}
\description{
The package provides an estimate of biological meaningful variance proportion for a target protein. This value is between 0 and 1, with 1 meaning the protein level is perfectly measured and 0 meaning the measurement is pure noise.
}
\author{
\strong{Maintainer}: Xilin Jiang \email{xj262@medschl.cam.ac.uk} (\href{https://orcid.org/0000-0001-6773-9182}{ORCID})

}
\keyword{internal}
