% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EWR_functions.R
\name{Factor_modelselection_COV_directML}
\alias{Factor_modelselection_COV_directML}
\title{Factor model selection.}
\usage{
Factor_modelselection_COV_directML(
  X,
  Y,
  Z,
  X_padding_num = 1,
  Y_padding_num = 1,
  max_XZ_factor = 5,
  max_Z_factor = 5,
  num_jk_block = 10
)
}
\arguments{
\item{X}{a vector which denote the levels of the target protein.}

\item{Y}{a vector (same length of X) which denote the levels of the helper protein, should be significantly correlated with X,}

\item{Z}{a matrix (each row correspond to one element in X and Y) of auxiliary proteins that tags biological variance in X. We recommend at least 100 columns.}

\item{X_padding_num}{default 1: increasing this number will putting more weights on target protein using a padding procedure. Usually not required as factor model is already controlling latent factors in Z.}

\item{Y_padding_num}{default 1: increasing this number will putting more weights on helper protein using a padding procedure. Usually not required as factor model is already controlling latent factors in Z.}

\item{max_XZ_factor}{default 5: grid search up to the maximum number of latent factors of X that are independent of Y.}

\item{max_Z_factor}{default 5: grid search up to the maximum number of latent factors of Z that are independent of X and Y.}

\item{num_jk_block}{default 10: jackknife blocks used to test which model explain more variance in X.}
}
\value{
a list containing objects:
\describe{
\item best_XZ: The choice of optimal latent factors of X that are independent of Y.
\item best_Z: The choice of optimal latent factors of Z that are independent of X and Y.
\item para_z_factors: comparison of metrics that are used to compare different models, we use sigma_X which denote the residual variance of X.
\item para_z_factors_jk: jackknife results for choosing best_Z.
\item  para_XZ_factors_jk: jackknife results for choosing best_XZ.
\item  summary_jackknife_testing_Z_factor: summary across jackknife blocks for choosing best_Z.
\item  summary_jackknife_testing_XZ: summary across jackknife blocks for choosing best_XZ.
}
}
\description{
\code{Factor_modelselection_COV_directML()} run different factor model with different number of latent factors.
The model select the model structure that maximise the variance explained in target protein. We use a
jackknifing procedure to test if adding a new latent factor explains more variance in the target protein.
}
\examples{
library(ExpressionWideRegression)
factor_model_selection <- Factor_modelselection_COV_directML(X_EXAMPLE, Y_EXAMPLE, Z_EXAMPLE,
 X_padding_num = 1,
 Y_padding_num = 1,
 max_XZ_factor = 3,
 max_Z_factor = 3,
 num_jk_block = 5)

}
