% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{X_EXAMPLE}
\alias{X_EXAMPLE}
\title{Example of target protein data}
\format{
\subsection{\code{X_EXAMPLE}}{

A vector of 5,000 elements.
}
}
\usage{
X_EXAMPLE
}
\description{
The data is generated by sampling procedure for showing the step-to-step application of the function.
}
\keyword{datasets}
