% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Y_EXAMPLE}
\alias{Y_EXAMPLE}
\title{Example of helper protein data}
\format{
\subsection{\code{Y_EXAMPLE}}{

A vector of 5,000 elements.
}
}
\usage{
Y_EXAMPLE
}
\description{
This helper protein is correlated with \code{X_EXAMPLE}.
}
\keyword{datasets}
