% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Z_EXAMPLE}
\alias{Z_EXAMPLE}
\title{Example of auxiliary protein data}
\format{
\subsection{\code{Z_EXAMPLE}}{

A matrix with 5,000 rows and 200 columns.
}
}
\usage{
Z_EXAMPLE
}
\description{
The data is generated by sampling procedure for showing the step-to-step application of the function.
}
\keyword{datasets}
